backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
asyncpg==0.30.0
pgvector==0.3.4
pydantic==2.10.4
pydantic-settings==2.6.1
python-dotenv==1.0.1
google-generativeai==0.8.3
langchain==0.3.13
langchain-google-genai==2.0.4
pymupdf==1.25.2
boto3==1.35.90
tenacity==9.0.0
python-multipart==0.0.20
alembic==1.13.1
httpx==0.28.0
beautifulsoup4==4.12.3
lxml==5.3.0
requests==2.32.3
